+++ tradeengine/engine/trade.go
@@ -5,7 +5,7 @@ import "encoding/json"
 type Trade struct {
 	TakerOrderID string `json:"taker_order_id"`
 	MakerOrderID string `json:"maker_order_id"`
-	Amount       uint64 `json:"amount"`
+	Amount       uint32 `json:"amount"`
 	Price        uint64 `json:"price"`
 }
